Simplify and reorganize Agent build for multi-file
We need a way to test the agent, and that means adding a websocket
client that can connect to the agent, which means adding a new class,
which means that we're outgrowing our rinky-dink index.ts

This moves the source files into the canonical `src/` directory, and
splits the server out into its own file. That way, we can move new
source files in as we will.

Also, this migrates away from `microbundle`, since the control classes
are not used anywhere except for in node and so there's no need to
build separate `umd` and `ejs` bundles. In fact, there's no need for
bundles at all.
