+++ packages/agent/src/index.ts
@@ -1,0 +1,1 @@
+export { AgentServer } from './server';
