+++ packages/server/package.json
@@ -36,7 +36,7 @@
     "@babel/core": "^7.0.0-0",
     "@babel/plugin-transform-runtime": "7.8.3",
     "@babel/preset-env": "7.8.4",
-    "@bigtest/agent": "^0.1.0-b37aeec",
+    "@bigtest/agent": "^0.1.0",
     "@types/express": "^4.17.1",
     "@types/graphql": "^14.5.0",
     "@types/http-proxy": "^1.17.0",
